requirements.txt
attrs>=18.2.0
azure-common>=1.1.18
azure-core>=1.24.0
azure-identity>=1.10.0
azure-mgmt-subscription>=3.0.0
beautifulsoup4>=4.0.0
bokeh>=1.4.0, <=3.1.0
cryptography>=3.1
deprecated>=1.2.4
dnspython>=2.0.0, <3.0.0
folium>=0.9.0
geoip2>=2.9.0
httpx==0.23.3
html5lib
ipython >= 7.1.1; python_version < "3.8"
ipython >= 7.23.1; python_version >= "3.8"
ipywidgets>=7.4.2, <8.0.0
KqlmagicCustom[jupyter-basic,auth_code_clipboard]>=0.1.114.post22
lxml>=4.6.5
matplotlib>=3.0.0
msal>=1.12.0
msal_extensions>=0.3.0
msrest>=0.6.0
msrestazure>=0.6.0
nest_asyncio>=1.4.0
networkx>=2.2
numpy>=1.15.4  # pandas
pandas>=1.4.0, <2.0.0
pygments>=2.0.0
pyjwt>=2.3.0
python-dateutil>=2.8.1  # pandas
pytz>=2019.2  # pandas
pyyaml>=3.13
setuptools>=40.6.3
tldextract>=2.2.2
tqdm>=4.36.1
typing-extensions>=4.2.0
urllib3>=1.23
